requirements.txt
fastapi
uvicorn[standard]
pdf2image
pillow
pytesseract
python-multipart
PyPDF2
PyMuPDF
